+++ views/motion/_view_prevnext.php
@@ -18,14 +18,17 @@ if (!$consultation->getSettings()->motionPrevNextLinks) {
 $prevMotion = null;
 $nextMotion = null;
 
+// Separate motions from resolutions for pagination
+$motionsOrResolutions = array_values(array_filter($consultation->motions, fn (Motion $itMotion) => $itMotion->isResolution() === $motion->isResolution()));
+
 $invisibleStatuses = $consultation->getStatuses()->getInvisibleMotionStatuses();
 if (in_array($motion->status, $invisibleStatuses) && User::havePrivilege($consultation, Privileges::PRIVILEGE_ANY, PrivilegeQueryContext::anyRestriction())) {
-    $motions = array_values(array_filter($consultation->motions, fn(Motion $motion) => in_array($motion->status, $invisibleStatuses)));
+    $motions = array_values(array_filter($motionsOrResolutions, fn(Motion $motion) => in_array($motion->status, $invisibleStatuses)));
     usort($motions, function(Motion $a, Motion $b) {
         return $a->getTimestamp() <=> $b->getTimestamp();
     });
 } else {
-    $motions = \app\components\MotionSorter::getSortedIMotionsFlat($consultation, $consultation->motions);
+    $motions = \app\components\MotionSorter::getSortedIMotionsFlat($consultation, $motionsOrResolutions);
 }
 foreach ($motions as $idx => $itMotion) {
     if ($motion->id !== $itMotion->id) {
@@ -43,6 +46,14 @@ if (!$nextMotion && !$prevMotion) {
     return;
 }
 
+if ($motion->isResolution()) {
+    $prevLabel = Yii::t('motion', 'prevnext_links_prev_res');
+    $nextLabel = Yii::t('motion', 'prevnext_links_next_res');
+} else {
+    $prevLabel = str_replace('%TYPE%', $motion->getMyMotionType()->titleSingular, Yii::t('motion', 'prevnext_links_prev'));
+    $nextLabel = str_replace('%TYPE%', $motion->getMyMotionType()->titleSingular, Yii::t('motion', 'prevnext_links_next'));
+}
+
 ?>
 <nav class="motionPrevNextLinks <?= ($top ? 'toolbarBelowTitle' : 'toolbarAtBottom') ?>">
     <?php
@@ -51,7 +62,7 @@ if (!$nextMotion && !$prevMotion) {
     <div class="prev">
         <a href="<?= Html::encode(UrlHelper::createIMotionUrl($prevMotion)) ?>">
             <span class="glyphicon glyphicon-chevron-left" aria-hidden="true"></span>
-            <?= str_replace('%TYPE%', $motion->getMyMotionType()->titleSingular, Yii::t('motion', 'prevnext_links_prev')) ?>
+            <?= $prevLabel ?>
         </a>
     </div>
     <?php
@@ -60,7 +71,7 @@ if (!$nextMotion && !$prevMotion) {
     ?>
     <div class="next">
         <a href="<?= Html::encode(UrlHelper::createIMotionUrl($nextMotion)) ?>">
-            <?= str_replace('%TYPE%', $motion->getMyMotionType()->titleSingular, Yii::t('motion', 'prevnext_links_next')) ?>
+            <?= $nextLabel ?>
             <span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span>
         </a>
     </div>